docs: update api sprint 1.5 (#206)
* fix(api): fix endpoint error

* fix(api): sync openapi with be codes

* docs(api): enhance OpenAPI Object and Components Object

* docs(api): apply common response

* refactor(be): allow nullable picture field

* feat(be): add profile dto validation

* refactor(be): rename Post.authorId to profileId

* refactor(be): use XXXResponse instead of FeedXXX

* feat(be): add URL validation

* docs(api): update required and nullable

* docs(api): add required to auth and remove unused res

* docs(api): revert status codes to 200

* chore(be): update CLAUDE.md and settings

* feat(be): add Recommendations objects

* docs(api): add Recommendations apis

* chore(be): rename Message.senderId to memberId

* feat(db): add MessageType enum

* refactor(be): change unreadCount type long to int

* feat(be): add Chat dtos

* docs(api): add chat apis

* feat(api): add pagination objects

* feat(api): add pagination objects

* docs(api): add feed, message pagination

* docs(api): remove unused endpoints

* docs(api): add checkUsername endpoint

* docs(api): add getMembers admin endpoint

* docs(api): add coworker request list endpoints

* docs(api): add getFeed endpoint

* docs(api): add getCoworkerTasks endpoint

* docs(api): add enum endpoint

* docs(api): add getTasks endpoint date filter

* docs(api): add MaskedMemberResponse

* feat(api): update dtos with MaskedMember

* docs(api): change ProfileResponse to ProfileWithMemberResponse

* docs(api): rename domain schema

* docs(api): add profile dto for detail

* docs(api): move (ADMIN) suffix to end of summary


* docs(api): rename ProfileWithMember to ProfileAndMember


* fix(api): use MaskedMemberResponse in FeedResponse


* docs(api): add profileId filter parameter to feed list API


---------

+++ apps/api/core/src/main/java/so/morton/api/domain/chat/Chat.java
@@ -10,11 +10,11 @@ public record Chat(
     String title,
     List<Long> participantIds,
     Message lastMessage,
-    long unreadCount,
+    int unreadCount,
     LocalDateTime createdAt,
     LocalDateTime modifiedAt
 ) {
-    public static Chat of(ChatEntity entity, List<Long> participantIds, Message lastMessage, long unreadCount) {
+    public static Chat of(ChatEntity entity, List<Long> participantIds, Message lastMessage, int unreadCount) {
         return new Chat(
                 entity.getId(),
                 entity.getTitle(),
